+++ mod-pubsub-server/src/main/java/org/folio/rest/impl/PubSubImpl.java
@@ -13,6 +13,7 @@ import org.folio.rest.jaxrs.model.Errors;
 import org.folio.rest.jaxrs.model.EventDescriptor;
 import org.folio.rest.jaxrs.model.MessagingModule.ModuleRole;
 import org.folio.rest.jaxrs.model.PublisherDescriptor;
+import org.folio.rest.jaxrs.model.SubscriberDescriptor;
 import org.folio.rest.jaxrs.resource.Pubsub;
 import org.folio.rest.tools.utils.TenantTool;
 import org.folio.rest.util.ExceptionHelper;
@@ -28,6 +29,7 @@ import java.util.Map;
 
 import static java.lang.String.format;
 import static org.folio.rest.jaxrs.model.MessagingModule.ModuleRole.PUBLISHER;
+import static org.folio.rest.jaxrs.model.MessagingModule.ModuleRole.SUBSCRIBER;
 
 public class PubSubImpl implements Pubsub {
 
@@ -152,29 +154,12 @@ public class PubSubImpl implements Pubsub {
     }
   }
 
-//  @Override
-//  public void deletePubsubEventTypesDeclarePublisher(String moduleName, String eventType, Map<String, String> okapiHeaders,
-//                                                     Handler<AsyncResult<Response>> asyncResultHandler, Context vertxContext) {
-//    try {
-//      MessagingModuleFilter filter = getMessagingModuleFilter(eventType, PUBLISHER, tenantId);
-//      messagingModuleService.deleteByModuleNameAndFilter(moduleName, filter)
-//        .map(DeletePubsubEventTypesDeclarePublisherResponse.respond204WithTextPlain(
-//          format("Publisher for event type '%s' was successfully deleted", eventType)))
-//        .map(Response.class::cast)
-//        .otherwise(ExceptionHelper::mapExceptionToResponse)
-//        .setHandler(asyncResultHandler);
-//    } catch (Exception e) {
-//      LOGGER.error("Failed to delete publisher", e);
-//      asyncResultHandler.handle(Future.succeededFuture(ExceptionHelper.mapExceptionToResponse(e)));
-//    }
-//  }
-
   @Override
-  public void deletePubsubEventTypesDeclarePublisherByEventTypeName(String eventTypeName, String moduleName, Map<String, String> okapiHeaders, Handler<AsyncResult<Response>> asyncResultHandler, Context vertxContext) {
+  public void deletePubsubEventTypesPublishersByEventTypeName(String eventTypeName, String moduleName, Map<String, String> okapiHeaders, Handler<AsyncResult<Response>> asyncResultHandler, Context vertxContext) {
     try {
       MessagingModuleFilter filter = getMessagingModuleFilter(eventTypeName, PUBLISHER, tenantId);
       messagingModuleService.deleteByModuleNameAndFilter(moduleName, filter)
-        .map(DeletePubsubEventTypesDeclarePublisherByEventTypeNameResponse.respond204WithTextPlain(
+        .map(DeletePubsubEventTypesPublishersByEventTypeNameResponse.respond204WithTextPlain(
           format("Publisher for event type '%s' was successfully deleted", eventTypeName)))
         .map(Response.class::cast)
         .otherwise(ExceptionHelper::mapExceptionToResponse)
@@ -195,6 +180,53 @@ public class PubSubImpl implements Pubsub {
         .setHandler(asyncResultHandler);
     } catch (Exception e) {
       LOGGER.error("Failed to get publishers by event type '{}'", e, eventTypeName);
+      asyncResultHandler.handle(Future.succeededFuture(ExceptionHelper.mapExceptionToResponse(e)));
+    }
+  }
+
+  @Override
+  public void postPubsubEventTypesDeclareSubscriber(SubscriberDescriptor entity, Map<String, String> okapiHeaders, Handler<AsyncResult<Response>> asyncResultHandler, Context vertxContext) {
+    try {
+      messagingModuleService.validateSubscriberDescriptor(entity)
+        .compose(errors -> errors.getTotalRecords() > 0
+          ? Future.succeededFuture(PostPubsubEventTypesDeclareSubscriberResponse.respond400WithApplicationJson(errors))
+          : messagingModuleService.saveSubscriber(entity, tenantId)
+            .map(v -> PostPubsubEventTypesDeclareSubscriberResponse.respond201()))
+        .map(Response.class::cast)
+        .otherwise(ExceptionHelper::mapExceptionToResponse)
+        .setHandler(asyncResultHandler);
+    } catch (Exception e) {
+      LOGGER.error("Failed to create subscriber", e);
+      asyncResultHandler.handle(Future.succeededFuture(ExceptionHelper.mapExceptionToResponse(e)));
+    }
+  }
+
+  @Override
+  public void getPubsubEventTypesSubscribersByEventTypeName(String eventTypeName, Map<String, String> okapiHeaders, Handler<AsyncResult<Response>> asyncResultHandler, Context vertxContext) {
+    try {
+      messagingModuleService.getByEventTypeAndRole(eventTypeName, SUBSCRIBER, tenantId)
+        .map(GetPubsubEventTypesSubscribersByEventTypeNameResponse::respond200WithApplicationJson)
+        .map(Response.class::cast)
+        .otherwise(ExceptionHelper::mapExceptionToResponse)
+        .setHandler(asyncResultHandler);
+    } catch (Exception e) {
+      LOGGER.error("Failed to get subscribers by event type '{}'", e, eventTypeName);
+      asyncResultHandler.handle(Future.succeededFuture(ExceptionHelper.mapExceptionToResponse(e)));
+    }
+  }
+
+  @Override
+  public void deletePubsubEventTypesSubscribersByEventTypeName(String eventTypeName, String moduleName, Map<String, String> okapiHeaders, Handler<AsyncResult<Response>> asyncResultHandler, Context vertxContext) {
+    try {
+      MessagingModuleFilter filter = getMessagingModuleFilter(eventTypeName, SUBSCRIBER, tenantId);
+      messagingModuleService.deleteByModuleNameAndFilter(moduleName, filter)
+        .map(DeletePubsubEventTypesSubscribersByEventTypeNameResponse.respond204WithTextPlain(
+          format("Subscriber for event type '%s' was successfully deleted", eventTypeName)))
+        .map(Response.class::cast)
+        .otherwise(ExceptionHelper::mapExceptionToResponse)
+        .setHandler(asyncResultHandler);
+    } catch (Exception e) {
+      LOGGER.error("Failed to delete subscriber", e);
       asyncResultHandler.handle(Future.succeededFuture(ExceptionHelper.mapExceptionToResponse(e)));
     }
   }